refactor: Server Component + 개별 mutation hooks 패턴으로 마이그레이션
- Services를 server/client 파일로 분리하여 빌드 에러 해결
- 4개 페이지를 Server Component로 전환 (Settings, Team, Category, Product)
- 13개 개별 mutation hooks 생성 및 deprecated hooks 제거
- 모든 사용자 대면 메시지 한글화
